requirements.txt
asgiref==3.8.1
branca==0.8.0
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
django-cors-headers==4.5.0
django-extensions==3.2.3
django-stubs==5.1.0
django-stubs-ext==5.1.0
djangorestframework==3.15.2
drf-yasg==1.21.8
folium==0.18.0
GDAL==3.6.2
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==2.1.2
packaging==24.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.1
types-PyYAML==6.0.12.20240917
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
xyzservices==2024.9.0